od6.4: Add input for user to change viewer precision in 2D viewer
Summary:
Jira OD-2276
The functionality for amplitude precision in 2D viewer has been added and tested
X and Y coordinates of 2D viewer now display till 2 decimal places instead of integers.
Z value precision has been corrected using Dynamic casting and passing point coordinates pt.y
All changes made for amplitude value have been removed
Enum has been removed and directly integers are passed to the user
Many keys have been introduced as static const chars

Test Plan:
Testing has been done on inl, crl and Z slices.
1. Checked before opening property dialog
2. Checked after opening dialog but not making any changes
3. Checked after making changes in annotation tab in dialog box
4. The number of decimals for the user to choose from has been connected to the survey sampling rate.
Now if there is a depth step of:
a) 10, user can choose to display upto 2 decimals with default 1
a) 0.05, user can choose to display upto 5 decimal with default 4

Reviewers: nageswara, nanne, arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D12917

+++ src/uiODMain/uiodviewer2d.cc
@@ -1023,7 +1023,7 @@ void uiODViewer2D::mouseMoveCB( CallBacker* cb )
     if ( valstr.isEmpty() ) valstr = pars.find( "Z-Coord" );
     if ( !valstr.isEmpty() )
     {
-	mousepos.z = valstr.toDouble() / zDomain().userFactor();
+	mousepos.z = valstr.toFloat() / zDomain().userFactor();
 	if ( datatransform_ )
 	    mousepos.z = datatransform_->transformBack( mousepos );
     }
